+++ source4/heimdal/lib/gssapi/mech/gss_aeap.c
@@ -1,6 +1,6 @@
 /*
  * AEAD support
- */ 
+ */
 
 #include "mech_locl.h"
 
@@ -90,7 +90,7 @@ gss_unwrap_iov(OM_uint32 *minor_status,
 	       int iov_count)
 {
 	struct _gss_context *ctx = (struct _gss_context *) context_handle;
-	gssapi_mech_interface m;	  
+	gssapi_mech_interface m;
 
 	if (minor_status)
 	    *minor_status = 0;
@@ -168,7 +168,7 @@ gss_release_iov_buffer(OM_uint32 *minor_status,
 		       int iov_count)
 {
     OM_uint32 junk;
-    size_t i;
+    int i;
 
     if (minor_status)
 	*minor_status = 0;
