+++ source4/heimdal/lib/roken/dumpdata.c
@@ -81,7 +81,7 @@ rk_undumpdata(const char *filename, void **buf, size_t *size)
     sret = net_read(fd, *buf, *size);
     if (sret < 0)
 	ret = errno;
-    else if (sret != *size) {
+    else if (sret != (ssize_t)*size) {
 	ret = EINVAL;
 	free(*buf);
 	*buf = NULL;
